requirements.txt
requests==2.32.5
pandas==3.0.1
snowflake-connector-python==4.3.0